#73 Fix : 깃허브 커밋 수 기준 주간으로 변경
