+++ src/main/java/com/devsong/server/ranking/controller/RankingController.java
@@ -1,7 +1,8 @@
-package com.devsong.server.ranking.BOJ.controller;
+package com.devsong.server.ranking.controller;
 
-import com.devsong.server.ranking.BOJ.dto.BojRankingResponseDto;
-import com.devsong.server.ranking.BOJ.service.RankingService;
+import com.devsong.server.ranking.dto.BojRankingResponseDto;
+import com.devsong.server.ranking.service.RankingService;
+import com.devsong.server.ranking.dto.GithubRankingResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -19,5 +20,10 @@ public class RankingController {
     @GetMapping("/boj")
     public List<BojRankingResponseDto> getBojRanking() {
         return rankingService.getBojRanking();
+    }
+
+    @GetMapping("/github")
+    public List<GithubRankingResponseDto> getGithubRanking() {
+        return rankingService.getGithubRanking();
     }
 }
