+++ src/main/java/com/devsong/server/ranking/dto/GithubSearchResponseDto.java
@@ -1,4 +1,4 @@
-package com.devsong.server.githubRanking.github.dto;
+package com.devsong.server.ranking.dto;
 
 import lombok.Data;
