--- src/main/java/com/devsong/server/ranking/github/controller/GithubRankingController.java
@@ -1,23 +1,0 @@
-package com.devsong.server.ranking.github.controller;
-
-import com.devsong.server.ranking.github.dto.GithubRankingResponseDto;
-import com.devsong.server.ranking.github.service.GithubRankingService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/ranking")
-public class GithubRankingController {
-
-    private final GithubRankingService rankingService;
-
-    @GetMapping("/github")
-    public List<GithubRankingResponseDto> getGithubRanking() {
-        return rankingService.getGithubRanking();
-    }
-}
