--- src/main/java/com/devsong/server/ranking/github/dto/GithubSearchResponseDto.java
@@ -1,9 +1,0 @@
-package com.devsong.server.ranking.github.dto;
-
-import lombok.Data;
-
-@Data
-public class GithubSearchResponseDto {
-    private int total_count; //커밋 수
-    private boolean incomplete_results; //전체 결과를 불러왔는지 여부
-}
